Temporary test changes to check reactions for DNA and test of voxel spac

+++ CellEngineProject/CellEngineTypes.h
@@ -9,8 +9,11 @@
 using SignedInt = std::int64_t;
 using UnsignedInt = std::uint64_t;
 
-using EntityIdInt = std::uint32_t;
-using ChainIdInt = std::uint32_t;
+//using EntityIdInt = std::uint32_t;
+//using ChainIdInt = std::uint32_t;
+
+using EntityIdInt = std::uint16_t;
+using ChainIdInt = std::uint16_t;
 
 template<class T>
 struct vector3
@@ -21,6 +24,7 @@ struct vector3
 using vector3_16 = vector3<uint16_t>;
 using vector3_64 = vector3<uint64_t>;
 
+//inline vector3_16 GetVector3FormVMathVec3(const vmath::vec3& Color)
 inline vector3_16 GetVector3FormVMathVec3(vmath::vec3 Color)
 {
     return {static_cast<uint16_t>(Color.X() * 100.00f), static_cast<uint16_t>(Color.Y() * 100.00f), static_cast<uint16_t>(Color.Z() * 100.00f) };
